fix: Revert inverted drive direction code
This commit should revert the changes made that made the robot drive in the opposite direction (These changes were originally made because black team's intake was on the back of the robot and they wanted the back to be the front). To achieve the same effect the motor IDs can be configure and the IMU can be recalibrated. Doing this will have the same effect without any of the bugs.