feat: Combine vision and non-vision IMU reset into one command
This commit combines both options for resetting the IMU into a single command. This gives the driver greater control over over the robot, and allows them to reset the IMU without ramming the robot into a wall, while still allowing for that method to be preformed if resetting the IMU with vision fails.